pyproject.toml
[tool.poetry]
name = "umapi-client"
version = "2.20"
description = "Client for the User Management API (UMAPI) from Adobe - see https://adobe.ly/2h1pHgV"
readme = "README.md"
authors = ["Andrew Dorton <adorton@adobe.com>", "Dan Brotsky", "Danimae Vossen", "Kevin Bhunut"]
maintainers = ["Andrew Dorton <adorton@adobe.com>"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
]
license = "MIT"
homepage = "https://github.com/adobe-apiplatform/umapi-client.py/"
documentation = "https://adobe-apiplatform.github.io/umapi-client.py/"
packages = [
    { include = "umapi_client" }
]

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.25.1"
cryptography = "3.3.2"
PyJWT = "2.4.0"

[tool.poetry.dev-dependencies]
mock = "^4.0.3"
pytest = "^6.2.3"
PyYAML = "^5.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
